+++ CHANGELOG.md
@@ -74,6 +74,7 @@ We’ve made fixes to GOV.UK Frontend in the following pull requests:
 - [#2807: Tidy up and refactor the Character Count JavaScript](https://github.com/alphagov/govuk-frontend/pull/2807)
 - [#2811: Use Element.id to get module id for accordion](https://github.com/alphagov/govuk-frontend/pull/2811)
 - [#2821: Avoid duplicated --error class on Character Count](https://github.com/alphagov/govuk-frontend/pull/2821)
+- [#2837: Target label active state as well as input focus on checkbox focus styles](https://github.com/alphagov/govuk-frontend/pull/2827). Thanks to [Ed Horsford](https://github.com/edwardhorsford) for raising and investigating this issue
 
 ## 4.3.1 (Patch release)